Added some visuals in the tilemap, also improved doors and added fog (it's a bit of a hack)
Easy stuff out the way first:

Added some actual tiles for the tilemap, mostly the floor but I made a shitty wall that I need to improve to, as well as making a table for the downstairs level.

Now the wacky stuff. I made doors openable from both ways, and I also split them into two different doors, horizontal doors and vertical doors.

This is because the doors open by manipulating the rotation of the door to make a swinging animation. This means that the doors can't be rotated in editor, or it'll mess with the animations. So now if you want a horizontal door, you add a horizontal door, same for vertical

Next, I added fog, so that you can only see areas you've opened. This is done by adding a fog object to the world, then adding a polygon2d and making the fog area you want.

After that, you add which doors you want to uncover that fog in the exported array in the editor.

It works really well but it's a bit messy and makes the scene tree a bit full. Sorry!